+++ spec/support/unit/helpers/action_pack_versions.rb
@@ -8,7 +8,11 @@ module UnitTests
     end
 
     def action_pack_gte_5?
-      action_pack_version =~ '>= 5'
+      action_pack_version >= 5
+    end
+
+    def action_pack_lt_5?
+      action_pack_version < 5
     end
 
     def action_pack_version