+++ spec/unit/shoulda/matchers/action_controller/callback_matcher_spec.rb
@@ -54,29 +54,29 @@ describe Shoulda::Matchers::ActionController::CallbackMatcher, type: :controller
     end
   end
 
-  describe '#use_before_filter' do
-    it_behaves_like 'CallbackMatcher', :before, :filter
-  end
-
-  describe '#use_after_filter' do
-    it_behaves_like 'CallbackMatcher', :after, :filter
-  end
-
-  describe '#use_around_filter' do
-    it_behaves_like 'CallbackMatcher', :around, :filter
-  end
-
-  if rails_4_x?
-    describe '#use_before_action' do
-      it_behaves_like 'CallbackMatcher', :before, :action
+  if action_pack_lt_5?
+    describe '#use_before_filter' do
+      it_behaves_like 'CallbackMatcher', :before, :filter
     end
 
-    describe '#use_after_action' do
-      it_behaves_like 'CallbackMatcher', :after, :action
+    describe '#use_after_filter' do
+      it_behaves_like 'CallbackMatcher', :after, :filter
     end
 
-    describe '#use_around_action' do
-      it_behaves_like 'CallbackMatcher', :around, :action
+    describe '#use_around_filter' do
+      it_behaves_like 'CallbackMatcher', :around, :filter
     end
+  end
+
+  describe '#use_before_action' do
+    it_behaves_like 'CallbackMatcher', :before, :action
+  end
+
+  describe '#use_after_action' do
+    it_behaves_like 'CallbackMatcher', :after, :action
+  end
+
+  describe '#use_around_action' do
+    it_behaves_like 'CallbackMatcher', :around, :action
   end
 end