+++ spec/unit/shoulda/matchers/active_model/allow_mass_assignment_of_matcher_spec.rb
@@ -1,7 +1,7 @@
 require 'unit_spec_helper'
 
 describe Shoulda::Matchers::ActiveModel::AllowMassAssignmentOfMatcher, type: :model do
-  if rails_lte_5?
+  if action_pack_lt_5?
     context '#description' do
       context 'without a role' do
         it 'includes the attribute name' do