+++ NEWS.md
@@ -12,6 +12,12 @@ is now:
 * Drop support for Rails 4.0 and 4.1 as well as Ruby 2.0 and 2.1, since they've
   been end-of-lifed. The gem now supports Ruby 2.2+ and Rails 4.2+.
 
+* `use_before_filter`, `use_after_filter`, and `use_around_filter` are no longer
+  usable when using shoulda-matchers under Rails 5.x, as the corresponding
+  controller callback don't exist there.
+
+  * *PR: [#1054]*
+
 ### Bug fixes
 
 * Fix association matchers when used under Rails 5.x so that they make use of
@@ -56,6 +62,7 @@ is now:
 [#964]: https://github.com/thoughtbot/shoulda-matchers/pulls/964
 [#917]: https://github.com/thoughtbot/shoulda-matchers/pulls/917
 [#867]: https://github.com/thoughtbot/shoulda-matchers/issues/867
+[#1054]: https://github.com/thoughtbot/shoulda-matchers/pulls/1054
 
 ### Improvements